+++ src/main/java/com/complexible/common/csv/generator/ValueGenerator.java
@@ -1,0 +1,7 @@
+package com.complexible.common.csv.generator;
+
+import org.openrdf.model.Value;
+
+public interface ValueGenerator<V extends Value> {
+    V generate(int rowIndex, String[] row);
+}
