+++ src/main/java/com/complexible/common/csv/provider/RowNumberProvider.java
@@ -1,0 +1,7 @@
+package com.complexible.common.csv.provider;
+
+public class RowNumberProvider extends ValueProvider {
+    protected String provideValue(int rowIndex, String[] row) {
+        return String.valueOf(rowIndex);
+    }
+}
